Модуль стора корзины
