+++ src/frontend/src/App.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="app">
-    <component :is="layout" :cart="cart">
-      <router-view :cart="cart" @addToCart="onAddToCart" />
+    <component :is="layout">
+      <router-view />
     </component>
   </div>
 </template>
@@ -12,19 +12,9 @@ import AppSidebarLayout from "@/layouts/AppSidebarLayout";
 export default {
   name: "App",
   components: { AppLayout, AppSidebarLayout },
-  data() {
-    return {
-      cart: [],
-    };
-  },
   computed: {
     layout() {
       return this.$route.meta.layout || "AppLayout";
-    },
-  },
-  methods: {
-    onAddToCart(item) {
-      this.cart.push(item);
     },
   },
 };
