+++ src/frontend/src/views/Index.vue
@@ -7,7 +7,7 @@
         <BuilderDoughSelector />
         <BuilderSizeSelector />
         <BuilderIngredientsSelector />
-        <BuilderPizzaView @addToCart="onAddToCart" />
+        <BuilderPizzaView />
       </div>
     </form>
   </main>
@@ -26,44 +26,6 @@ export default {
     BuilderIngredientsSelector,
     BuilderSizeSelector,
     BuilderDoughSelector,
-  },
-  props: {
-    cart: {
-      type: Array,
-      required: true,
-    },
-  },
-  methods: {
-    onAddToCart() {
-      const pizzaIngredients = this.ingredients
-        .filter((item) => item.amount > 0)
-        .map((item) => {
-          return {
-            id: item.id,
-            amount: item.amount,
-            price: item.price,
-          };
-        });
-
-      const pizza = {
-        name: this.pizzaName,
-        dough: {
-          id: this.selectedDough.id,
-          price: this.selectedDough.price,
-        },
-        size: {
-          id: this.selectedSize.id,
-          multiplier: this.selectedSize.multiplier,
-        },
-        sauce: {
-          id: this.selectedSauce.id,
-          price: this.selectedSauce.price,
-        },
-        ingredients: pizzaIngredients,
-      };
-      this.$emit("addToCart", pizza);
-      this.$store.dispatch("Builder/resetState");
-    },
   },
 };
 </script>
